feat: order table lock handlings, point format across several tables, pwa attempt
module: config point settings
- modified config points settings: point calculation's point field to allow decimal points

module: order
- fixed process payment flow: customers' points round off to 2 decimal places
- added locked_by column to tables table to track the user who locked the table (only stockie admin users)
- modify flow to include adding locked_by user id when locking table and unlocking table
- added handling for when there are multiple tabs opened and on the ordertables page, opening a table to view the order/close drawer/refresh/redirect will update the unified locked tables array through the use of broadcast channel and session storage
- modified get all zone function to a post request and update table that are not viewed by the current auth user to be unlocked

module: product
- fixed create/edit product form: allow points field to have 2 decimal places

module: customer
- fixed customer - points adjustment form: allow points field to have 2 decimal places

module: transaction listing
- fixed refund transaction: format of refund points when processing the refund

module: loyalty programme
- fixed loyalty programme - add/edit tier form: allow bonus points field to have 2 decimal places

others:
- added cronjob to reset locked tables for users that doesnt have an active session
- added implementation of installing the website as a progressive web application (pwa) using vite pwa plugin

+++ database/migrations/2024_06_04_083309_create_point_histories_table.php
@@ -18,10 +18,10 @@ return new class extends Migration
             $table->string('type');
             $table->string('point_type');
             $table->integer('qty');
-            $table->integer('amount');
-            $table->integer('old_balance');
-            $table->integer('new_balance');
-            $table->integer('expire_balance')->nullable()->default(0);
+            $table->decimal('amount', 13, 2);
+            $table->decimal('old_balance', 13, 2);
+            $table->decimal('new_balance', 13, 2);
+            $table->decimal('expire_balance', 13, 2)->nullable()->default(0.00);
             $table->dateTime('expired_at')->nullable()->default(NULL);
             $table->string('remark')->nullable()->default(NULL);
             $table->unsignedBigInteger('customer_id');
